server/requirements.txt
Django~=4.2.2
django-multiselectfield==4.6.3